Move parse_pdb out to it's own module
