+++ pyse/parse_foldx_jobs.py
@@ -26,6 +26,8 @@ import sys
 from collections import defaultdict
 from multiprocessing import Pool, cpu_count
 
+from pdb import parse_pdb
+
 # Define and error printing function in lieu of logging api
 eprint = functools.partial(print, file=sys.stderr, flush=True)
 
@@ -121,29 +123,6 @@ def parse_raw_buildmodel(pdb, rawmodel):
     assert len(energies) == len(wt_energies)
 
     return energies, wt_energies
-
-def parse_pdb(pdb):
-    """ Read the x,y,z positions for each atom in a PDB file """
-    positions = []
-
-    def stripstr(v):
-        return str(v).strip()
-    for line in pdb:
-        if not line.startswith('ATOM'):
-            continue
-        # these are specified as inclusive ranges of 1-based
-        # indexes to match the PDB spec
-        fieldspec = [
-            ('atom', 13, 16, stripstr),
-            ('remnant', 18, 20, stripstr),
-            ('chain', 22, 22, stripstr),
-            ('position', 23, 26, int),
-            ('x', 31, 38, float),
-            ('y', 39, 46, float),
-            ('z', 47, 54, float),
-        ]
-        positions.append({f[0]: f[3](line[f[1] - 1:f[2]]) for f in fieldspec})
-    return positions
 
 def calculate_energy_deltas(energies, wt_energies):
     """Sum the columns in the energies and wt_energies lists and
